refactor: change all icons filenames to have ‘magic-sync’ fix
